+++ source/common/tables/dbo.catalogDatabaseNames.sql
@@ -29,12 +29,7 @@ CREATE TABLE [dbo].[catalogDatabaseNames]
 	(
 		[id],
 		[instance_id]
-	) ON [PRIMARY],
-	CONSTRAINT [UK_catalogDatabaseNames_Name] UNIQUE  NONCLUSTERED 
-	(
-		[name],
-		[instance_id]
-	) ON [PRIMARY],
+	) ON [PRIMARY]
 	CONSTRAINT [FK_catalogDatabaseNames_catalogProjects] FOREIGN KEY 
 	(
 		[project_id]
@@ -60,4 +55,5 @@ CREATE INDEX [IX_catalogDatabaseNames_InstanceID] ON [dbo].[catalogDatabaseNames
 GO
 CREATE INDEX [IX_catalogDatabaseNames_ProjecteID] ON [dbo].[catalogDatabaseNames]([project_id]) ON [PRIMARY]
 GO
-
+CREATE UNIQUE INDEX [UK_catalogDatabaseNames_Name] ON [dbo].[catalogDatabaseNames]([name], [instance_id]) INCLUDE ([project_id], [active])
+GO