+++ source/monitoring/stored-procedures/dbo.usp_monAlarmCustomTransactionsStatus.sql
@@ -1022,8 +1022,7 @@ WHILE @@FETCH_STATUS=0
 		/* check for additional receipients for the alert */		
 		SET @additionalRecipients = [dbo].[ufn_monGetAdditionalAlertRecipients](@projectID, @instanceName, @eventName, @databaseName)
 
-		EXEC [dbo].[usp_logEventMessageAndSendEmail]	@projectCode			= @projectCode,
-														@sqlServerName			= @instanceName,
+		EXEC [dbo].[usp_logEventMessageAndSendEmail]	@sqlServerName			= @instanceName,
 														@dbName					= @databaseName,
 														@objectName				= @severity,
 														@childObjectName		= @childObjectName,