+++ source/patches/20180713-patch-upgrade-from-v2018_6-to-v2018_7-common.sql
@@ -1,0 +1,19 @@
+SET NOCOUNT ON
+/*
+*-----------------------------------------------------------------------------*
+* dbaTDPMon (Troubleshoot Database Performance / Monitoring)                  *
+* https://github.com/rentadba/dbaTDPMon, under GNU (GPLv3) licence model      *
+*-----------------------------------------------------------------------------*
+* Patch script: from version 2018.6 to 2018.7 (2018.07.13)					  *
+*-----------------------------------------------------------------------------*
+*/
+
+/*---------------------------------------------------------------------------------------------------------------------*/
+/* patch module: common																								   */
+/*---------------------------------------------------------------------------------------------------------------------*/
+RAISERROR('* Patch: 20180713-patch-upgrade-from-v2018_6-to-v2018_7-common.sql', 10, 1) WITH NOWAIT
+
+IF NOT EXISTS(SELECT * FROM [dbo].[appConfigurations] WHERE [name] = 'Maximum SQL Agent jobs running on the same physical volume (0=unlimited)' and [module] = 'common')
+	INSERT	INTO [dbo].[appConfigurations] ([module], [name], [value])
+			SELECT 'common' AS [module], 'Maximum SQL Agent jobs running on the same physical volume (0=unlimited)'	AS [name], '0' AS [value]
+GO